requirements.txt
Django==3.2.12
django-cors-headers==3.10.0
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.3.0
psycopg2==2.9.9
psycopg2-binary==2.9.1